Fixed extremely annoying bug
This bug kept me up for way too many nights, including this one where I've finally squashed it at 3am. Never forget: just because your stuff is working right now doesn't mean you haven't written some bugs.

This bug was annoying because it was intermittant. It was annoying because it only popped up in later AE versions, and usually only in Release builds. It was annoing because it only mostly appeared in later versions of OS X, making me think it was an SDK issue. It was annoying because rather than a crash that would call attention to itself, the locked handles floated around until they produced errors which then wrecked other havoc. The bug was annoying because it was actually written 6 months ago. It was annoying because I was resizing handles in the Cryptomatte plug-in so I thought the bug was there, not in the trusty old  OpenEXR plug-in. (Not that the AE error messages told me which plug-in was to blame.)

Bugs are annoying. This one was my fault though. As a great man once said, "I hate computers."